requirements.txt
reportlab
Pillow
requests
beautifulsoup4  # needed for enrich_metadata.py
